refactor: store and hook
